+++ src/os/eeiop/cdvd/eecdvd.h
@@ -13,7 +13,7 @@ typedef struct {
 void CdvdInit();
 void CdvdInitSoftReset();
 int LoadReq(int file_no, uint64_t addr);
-int64_t LoadReqGetAddr(int file_no, uint64_t addr, int64_t *id);
+int64_t LoadReqGetAddr(int file_no, uint64_t addr, int *id);
 int LoadReqSe(int file_no, u_char se_type);
 int64_t LoadReqNSector(int file_no, int sector, int size, int64_t addr);
 int LoadReqNFno(int file_no, int64_t addr);
